move media embed to UI dir
